Refactor analysis modules with shared components and remove category analysis
- Extract shared components for income/expense analysis:
  - CategoryDetailList: Collapsible detail list with progress bars
  - TopTransactionsTable: Top 50 transactions table
  - MonthlyTrendChart: Monthly trend area chart
  - AccountDistributionChart: Account distribution bar chart
  - CategoryDistributionChart: Category distribution with filter
- Create useCategoryData hook for unified data processing
- Add category-shared types
- Remove category analysis module completely:
  - Delete CategoryPieChart, CategoryBreakdown, ExpenseRadar components
  - Remove from navigation and Index.tsx
- Fix tooltip colors to match income/expense themes
- Fix progress bar alignment in detail lists
