+++ src/components/dashboard/shared/index.ts
@@ -1,0 +1,11 @@
+export { CategoryDetailList } from './CategoryDetailList';
+export { TopTransactionsTable } from './TopTransactionsTable';
+export { MonthlyTrendChart } from './MonthlyTrendChart';
+export { AccountDistributionChart } from './AccountDistributionChart';
+export { CategoryDistributionChart } from './CategoryDistributionChart';
+
+export type { CategoryDetailListProps } from './CategoryDetailList';
+export type { TopTransactionsTableProps } from './TopTransactionsTable';
+export type { MonthlyTrendChartProps } from './MonthlyTrendChart';
+export type { AccountDistributionChartProps } from './AccountDistributionChart';
+export type { CategoryDistributionChartProps } from './CategoryDistributionChart';
